README.md updates: updates on pg_dump/restore.
There have been mentions that the --inserts method and COPY method via
pg_dump and pg_restore have been giving issues. Individuals importing
large amount of history data are stating that the restore method is
taking a very long time. The documentation has been updated to address
that keeping history numeric float data longer than a certain amount of
days is not the best practice for the Zabbix history tables. After a
specific amount of time the history values are used by the trends, a
built-in history data mechanism that uses the history values to display
information in Zabbix.

In addition notes regarding using the --inserts and COPY method of the
pg_dump and pg_restore were updated as both methods worked from personal
experience.